v 1.4

+++ tasklist.php
@@ -7,7 +7,7 @@ if(!isset($_SESSION["user_id"]))
 {
     header("location:index.php");
 }
-$id = $_SESSION["user_id"];
+$user_id = $_SESSION["user_id"];
 
 /**********************************************************/
                 /*Delete All Tasks */
@@ -34,16 +34,16 @@ if(isset($_POST['ready']))
     if($_POST['status'] == 0)
     {
         $id = $_POST['Id'];
-        $data = array('id' => $id);
-        $query = $db -> prepare("UPDATE tasks SET status = 1 WHERE id = :id");
+        $data = array('id' => $id,'user_id' => $user_id);
+        $query = $db -> prepare("UPDATE tasks SET status = 1 WHERE id = :id and user_id = :user_id");
         $query->execute($data);
         header('location: tasklist.php');
     }
     if($_POST['status'] == 1)
     {
         $id = $_POST['Id'];
-        $data = array('id' => $id);
-        $query = $db -> prepare("UPDATE tasks SET status = 0 WHERE id = :id");
+        $data = array('id' => $id,'user_id' => $user_id);
+        $query = $db -> prepare("UPDATE tasks SET status = 0 WHERE id = :id and user_id = :user_id");
         $query->execute($data);
         header('location: tasklist.php');
     }
@@ -51,8 +51,8 @@ if(isset($_POST['ready']))
 if(isset($_POST['delete']))
 {
     $id = $_POST['Id'];
-    $data = array('id' => $id);
-    $query = $db -> prepare("DELETE FROM tasks  WHERE id = :id");
+    $data = array('id' => $id,'user_id' => $user_id);
+    $query = $db -> prepare("DELETE FROM tasks  WHERE id = :id and user_id = :user_id");
     $query->execute($data);
     header('location: tasklist.php');
 }
@@ -62,53 +62,22 @@ if(isset($_POST['AddTask']))
 {
     $date = date('Y-m-d');
     $NewTask = htmlentities($_POST['NewTask']);
-    $data = array('id' => $id,'NewTask' => $NewTask,'date' => $date);
-    $query = $db -> prepare("INSERT INTO tasks (user_id,description,created_at) VALUES (:id,:NewTask,:date)");
+    $data = array('user_id' => $user_id,'NewTask' => $NewTask,'date' => $date);
+    $query = $db -> prepare("INSERT INTO tasks (user_id,description,created_at) VALUES (:user_id,:NewTask,:date)");
     $query->execute($data);
     header('location: tasklist.php');
 }
 /**********************************************************/
         /* Create function for create Tasks */
-function CreateTasks($id,$db)
+function CreateTasks($user_id,$db)
 {
-$data = array('id' => $id);
-$query = $db -> prepare("SELECT description,status,id FROM tasks Where '$id' = user_id");
+$data = array('user_id' => $user_id);
+$query = $db ->prepare("SELECT description,status,id FROM tasks Where $user_id = :user_id");
 $query->execute($data);
-if($query-> rowCount() > 0)
-    {
-        while ($row = $query->fetch(PDO::FETCH_ASSOC)) 
-        {
-            if($row['status'] == 1)
-            {
-                $ready = "Unready";
-            }
-            if($row['status'] == 0)
-            {
-                $ready = "Ready";
-            }
-            $task_id = $row['id'];
-            $status = $row['status'];
-            echo "<form action=\"tasklist.php\" method=\"POST\" name=\"ChangeTask\">";
-            echo $row['description'];
-            echo "<p></p>";
-            echo "<input type =\"submit\" value=\"$ready\" name=\"ready\">";
-            echo "<input type=\"submit\" value=\"Delete\" name=\"delete\">";
-            echo "<input type=\"hidden\" name=\"Id\" value=\"$task_id\"/>";
-            echo "<input type=\"hidden\" name=\"status\" value=\"$status\"/>";
-            if($row['status'] == 1)
-            {
-                echo "<img src=\"green_circle.png\">";
-            }
-            if($row['status'] == 0)
-            {
-                echo "<img src=\"red_circle.png\">";
-            }
-            echo "<p></p>";
-            echo "</form>";
-        
-        }
-        
-    }
+if($query->rowCount() > 0)
+   {
+       return $query;
+   }
 }
 ?>
 <!----------------------------------------------------->
@@ -136,5 +105,38 @@ if($query-> rowCount() > 0)
             <!-- HTML CODE -->
             
 <?php
-CreateTasks($id,$db);
+$query = CreateTasks($user_id,$db);
+if(isset($query))
+{
+while ($row = $query->fetch(PDO::FETCH_ASSOC)) 
+    {
+        if($row['status'] == 1)
+        {
+            $ready = "Unready";
+        }
+        if($row['status'] == 0)
+        {
+            $ready = "Ready";
+        }
+        $task_id = $row['id'];
+        $status = $row['status'];
+        echo "<form action=\"tasklist.php\" method=\"POST\" name=\"ChangeTask\">";
+        echo $row['description'];
+        echo "<p></p>";
+        echo "<input type =\"submit\" value=\"$ready\" name=\"ready\">";
+        echo "<input type=\"submit\" value=\"Delete\" name=\"delete\">";
+        echo "<input type=\"hidden\" name=\"Id\" value=\"$task_id\"/>";
+        echo "<input type=\"hidden\" name=\"status\" value=\"$status\"/>";
+        if($row['status'] == 1)
+        {
+            echo "<img src=\"green_circle.png\">";
+        }
+        if($row['status'] == 0)
+        {
+            echo "<img src=\"red_circle.png\">";
+        }
+        echo "<p></p>";
+        echo "</form>";     
+    }
+}
 ?>
